Created a function that returns a null job
Don't know if this is the right approach, but when converting
the JobUpdatePage to tsx, useState wouldn't work correctly
without having a Job object, and I don't know if fetching
the job before setting the state and rendering anything is
a good idea. I may refactor later so that the job or jobid
is passed to the component instead of getting the ID from the
address bar.